+++ packages/shared/friends/utils.ts
@@ -1,10 +1,3 @@
-import { put } from 'redux-saga/effects'
-import { updatePrivateMessagingState } from './actions'
-import { findPrivateMessagingFriendsByUserId, getPrivateMessaging } from './selectors'
-import { FriendsState, SocialData } from './types'
-import { store } from 'shared/store/isolatedStore'
-import { Conversation, SocialAPI } from 'dcl-social-client'
-
 /**
  * Get the local part of the userId from matrixUserId
  * @param userId a string with the matrixUserId pattern
@@ -22,29 +15,6 @@ export function getUserIdFromMatrix(userId: string) {
 }
 
 /**
- *
- */
-export async function getConversationId(client: SocialAPI, userId: string) {
-  let conversationId = findPrivateMessagingFriendsByUserId(store.getState(), userId)?.conversationId
-
-  if (!conversationId) {
-    const socialId = getMatrixIdFromUser(userId)
-    const conversation: Conversation = await client.createDirectConversation(socialId)
-
-    const socialData: SocialData = {
-      userId: userId,
-      socialId: socialId,
-      conversationId: conversation.id
-    }
-
-    updateSocialInfo(socialData)
-    conversationId = conversation.id
-  }
-
-  return conversationId
-}
-
-/**
  * Get the matrixUserId from userId
  * @param userId a string with the userId pattern
  *
@@ -52,27 +22,7 @@ export async function getConversationId(client: SocialAPI, userId: string) {
  * from: '0x1111ada11111'
  * to: '@0x1111ada11111:decentraland.org'
  * */
-function getMatrixIdFromUser(userId: string) {
+export function getMatrixIdFromUser(userId: string) {
   // TODO check how'd be the correct way to do this
   return '@' + userId.toString() + ':decentraland.org'
-}
-
-/**
- *
- */
-function updateSocialInfo(socialData: SocialData) {
-  const friends: FriendsState = getPrivateMessaging(store.getState())
-
-  // add social info
-  friends.socialInfo[socialData.socialId] = socialData
-
-  put(
-    updatePrivateMessagingState({
-      client: friends.client,
-      socialInfo: friends.socialInfo,
-      friends: friends.friends,
-      fromFriendRequests: friends.fromFriendRequests,
-      toFriendRequests: friends.toFriendRequests
-    })
-  )
 }
